move test_config and make_tuf_repo to TestBoards
